DAOS-7214 test: Vagrantfile to stand up a cluster
That can run DAOS.

Allow the REMOTE_ACCT to be specifed to help in testing outside of
Jenkins.

Start to set reasonable defaults for environment variables so that fewer
variables need to be specified when running manually.

Don't mount tmpfs on nodes with NVMe disk(s).

Add --nostdin to clush commands:
- due to https://github.com/cea-hpc/clustershell/issues/461

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>